+++ src/wasm-lib/kcl/Cargo.toml
@@ -18,8 +18,8 @@ chrono = "0.4.35"
 clap = { version = "4.5.2", features = ["cargo", "derive", "env", "unicode"], optional = true }
 dashmap = "5.5.3"
 databake = { version = "0.1.7", features = ["derive"] }
-derive-docs = { version = "0.1.10" }
-#derive-docs = { path = "../derive-docs" }
+#derive-docs = { version = "0.1.11" }
+derive-docs = { path = "../derive-docs" }
 futures = { version = "0.3.30" }
 gltf-json = "1.4.0"
 kittycad = { workspace = true }
